requirements.txt
click==7.1.2
Flask==1.1.2
Flask-MySQL==1.5.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysqlclient==2.0.3
PyMySQL==1.0.2
python-dotenv==0.17.1
Werkzeug==1.0.1
